Feat/catalog product modules

+++ src/products/products.service.ts
@@ -32,6 +32,15 @@ export class ProductsService {
   }
 
   async create(createProductDto: CreateProductDto): Promise<ProductEntity> {
+    const alreadyExists = await this.productRepo.findOne({
+      where: { name: createProductDto.name },
+    });
+
+    if (alreadyExists) {
+      throw new ConflictException(
+        `Product with name "${createProductDto.name}" already exists`,
+      );
+    }
     const product = this.productRepo.create(createProductDto);
     return await this.productRepo.save(product);
   }
@@ -97,12 +106,19 @@ export class ProductsService {
       throw new NotFoundException(`Product with ID: ${productId} not found`);
     }
 
+    const catalog = await this.catalogRepo.findOne({
+      where: { id: catalogId },
+    });
+    if (!catalog) {
+      throw new NotFoundException(`Catalog with ID: ${catalogId} not found`);
+    }
+
     const alreadyAssigned = product.catalogs.find(
       (catalog) => catalog.id === catalogId,
     );
 
     if (!alreadyAssigned) {
-      throw new ConflictException(
+      throw new NotFoundException(
         `Cannot Delete from Catalog: Product ID: ${productId} is not assigned to Catalog ID: ${catalogId}`,
       );
     }
